Moved `getExtent` to CoordinateUtils
